Simplify Job.get_runnable_tasks to only return tasks
Move the optional next_check datetime from being returned by
get_runnable_tasks always, to being a property of the Job object.

This simplifies implementing a new get_runnable_tasks method for
a job, as you only have to think about tasks, and if you do need
to control timing, you can implement that as part of a property.